refactor: relocate web/ to src/antennaknobs/web (namespaced subpackage)
The web API server lived at the repo root as a top-level `web` package,
which setup.py shipped into the wheel — dropping a dangerously generic
`web` module into site-packages where it can collide with anything. Move
the whole tree under the package so it imports as `antennaknobs.web`.

- git-mv web/ -> src/antennaknobs/web/ (all internal imports were already
  relative, so the tree moves as a self-consistent unit; vite outDir
  ../static and server.py's __file__-relative _FRONTEND_DIR both follow).
- setup.py: find_packages(where="src") now discovers antennaknobs.web[.*]
  automatically; package_dir={"":"src"}; package_data keyed on
  antennaknobs.web. No more explicit top-level `web` mapping.
- Rewrite the absolute `from web.X` / `import web.X` imports in the
  antennaknobs test suite to `antennaknobs.web.X` (the momwire submodule
  keeps its own independent `web/`, untouched).
- Update .gitignore, README (uvicorn antennaknobs.web.server:app, frontend
  path), CI frontend-build working-directory, and assorted comments.

Side benefit: web/ now sits under src/, so `coverage --source=src/`
measures it too. Verified: 915 tests pass; frontend builds to the new
static dir; TestClient serves /healthz, /examples, and the SPA at /.
