+++ README.md
@@ -55,7 +55,7 @@ so one process serves the whole app:
 
 ```bash
 pip install "antennaknobs[web]"
-uvicorn web.server:app                   # open http://127.0.0.1:8000
+uvicorn antennaknobs.web.server:app      # open http://127.0.0.1:8000
 ```
 
 The backend serves the UI at `/` and the JSON/`/ws` API on the same origin;
@@ -67,10 +67,10 @@ Vite dev server alongside the backend so you get HMR:
 ```bash
 # Terminal 1 — backend (from the repo root, in your .venv)
 pip install -e ".[web]"
-uvicorn web.server:app --reload          # API on http://127.0.0.1:8000
+uvicorn antennaknobs.web.server:app --reload   # API on http://127.0.0.1:8000
 
 # Terminal 2 — frontend dev server
-cd web/frontend
+cd src/antennaknobs/web/frontend
 npm install
 npm run dev                              # open http://localhost:5173
 ```
@@ -78,7 +78,7 @@ npm run dev                              # open http://localhost:5173
 The Vite dev server proxies the API and the `/ws` live-solve channel to the
 backend on port 8000, so you only ever open `http://localhost:5173`. (A source
 checkout has no pre-built bundle, so the backend alone runs API-only until you
-`npm run build` — which writes `web/static/`, the same bundle the wheel ships.)
+`npm run build` — which writes `src/antennaknobs/web/static/`, the same bundle the wheel ships.)
 
 > The `[web]` extra pulls in `uvicorn[standard]`, which includes the WebSocket
 > support the live-solve channel needs — plain `uvicorn` fails the `/ws`
