+++ pyproject.toml
@@ -41,16 +41,16 @@ dependencies = [
 ]
 
 [project.optional-dependencies]
-# Web UI: web/server.py is a FastAPI app served by uvicorn.
+# Web UI: src/antennaknobs/web/server.py is a FastAPI app served by uvicorn.
 #   - fastapi (+ its starlette core) is the web framework itself — imported
-#     directly by web/server.py, so it must be declared here, not assumed.
+#     directly by antennaknobs/web/server.py, so it must be declared here.
 #   - uvicorn[standard] is the ASGI server. The live-solve channel at /ws is a
 #     WebSocket upgrade, so the server needs WS support: bare `uvicorn` ships
 #     without it and fails the handshake at runtime. `uvicorn[standard]` already
 #     pulls in websockets (+ httptools + uvloop, uvloop a no-op on Windows), but
 #     we also list `websockets` explicitly so the /ws dependency is visible and
 #     pinned even if uvicorn's extra ever changes.
-#   - psutil powers _physical_cpu_count() in web.server — portable across
+#   - psutil powers _physical_cpu_count() in antennaknobs.web.server — portable across
 #     Windows/macOS/Linux. Falls back to os.cpu_count() if missing, which
 #     misfires on chips without HT (pins to half the actual cores).
 web = ["fastapi", "uvicorn[standard]", "websockets", "psutil"]
