+++ src/antennaknobs/web/server.py
@@ -9,7 +9,7 @@ The response shape is uniform across geometries — each wire is a sequence of
 knots with per-knot complex currents and the feed lives on one of the wires —
 so the frontend draws every geometry the same way.
 
-Run: uvicorn web.server:app --reload
+Run: uvicorn antennaknobs.web.server:app --reload
 (needs uvicorn[standard] — /ws is a WebSocket upgrade.)
 """
